case_beanalytic/requirements.txt
selenium==4.11.2
pandas==2.1.1
google-cloud-bigquery==3.11.0
